+++ api/src/main/java/com/ripstech/apiconnector2/entity/receive/application/Scan.java
@@ -13,8 +13,8 @@ import java.util.List;
 
 public class Scan {
 
-	private int id;
-	private Integer loc;
+	private long id;
+	private Long loc;
 	private Integer analysisDepth;
 	private Boolean fullCodeCompared;
 	private Upload upload;
@@ -40,11 +40,11 @@ public class Scan {
 	private String path;
 	private Integer maxIssuesPerType;
 
-	public int getId() {
+	public long getId() {
 		return this.id;
 	}
 
-	public Integer getLoc() {
+	public Long getLoc() {
 		return this.loc;
 	}
