+++ api/src/main/java/com/ripstech/apiconnector2/service/application/scan/issue/ReviewService.java
@@ -9,11 +9,11 @@ import java.util.List;
 
 public class ReviewService extends PostGetService<Review, ReviewSend> {
 
-	private final int applicationId;
-	private final int scanId;
-	private final int issueId;
+	private final long applicationId;
+	private final long scanId;
+	private final long issueId;
 
-	public ReviewService(String baseUri, int applicationId, int scanId, int issueId) {
+	public ReviewService(String baseUri, long applicationId, long scanId, long issueId) {
 		super(baseUri);
 		this.applicationId = applicationId;
 		this.scanId = scanId;
